requirements.txt
python-dotenv==1.0.1
Requests==2.31.0
